move sgx-specific calls to sgx directory.
